Converted moments to load incrementally
In order to display the correct number of assets in each moment, we need to perform a fetch (as per the documentation). This can take quite a bit of time on a large library. Instead, we now load the sections in 50 at a time. This also allows us to completely omit empty moments instead of trying to hide the header dynamically.

Tangentally, the moments view controller is now inverted so that we don't need to manually scroll to the bottom.