Fixed asset detail view
When we broke apart the collection view controllers we broke the detail screen.

The detail screen doesn't need to worry about moments because it just shows one asset at a time, so we can just fetch all assets (which represents the same assets as the moments as a single fetch result). This greatly simplifies navigation.