Fixed crash related to unauthorized access
Now the photo picker will check if it has access to the photo library and if it doesn't it will show a message to the user about why they can't see their photos.

This fixes a crash where a photo manager would be created even when the app didn't have access and then when it was deallocated it would crash the app.