Fix warning with Xcode 12
When using Xcode 12, there is a warning `The iOS Simulator deployment target 'IPHONEOS_DEPLOYMENT_TARGET' is set to 8.0, but the range of supported deployment target versions is 9.0 to 14.2.99.`. In order to get rid of that warning, the deployment target was bumped to iOS 9, but to maintain support for older versions of Xcode with a lower deployment target, the previous package is also included for older versions of Swift.